Move __assert_fail and strcpy into selfe-sys
Exporting these from rust libraries is unreliable. The previously present
re-exporting from selfe-runtime to libsel4-start is particularly bad, as the
ambiguity between the two versions leads to the compiler generating a recursive
function that eats the stack.

This implements the two functions in plain C, and puts them right next to the
code that needs them.